Fix bug where tdiff exited with code 2 without printing anything.
That occurs when missing files comparisons are disabled (--no-dirs)
and there are missing files.  Same for hardlinks.

* str_list.c (str_list_compare): When missing entries are
detected, delegate signaling an error to the reportMissing
callback, which must now return an int.

* str_list.h: Update str_list_compare prototype.

* tdiff.c (reportMissingXattr, reportMissingAcl)
(reportMissingFile, reportMissingHardLink): Return if the value
was found different or not (a discrepancy may not need to be
reported if a toggle option turn it off).

* tests/hardlinks.out.gold (tdiff):
* tests/directories.out.gold (tdiff):
Update, exit code should be 2.